Removed need for Enabled entry in config
The issue was that mods could be set to enabled in the SaveProfile definitions, but disabled in their own config files.
This, along with changes in their Plugin files, will have an Enabled entry in the default BepInEx config file, but won't have that entry in each SaveProfile config file.